requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiohttp==3.9.1
redis==5.0.1
apscheduler==3.10.4
pyyaml==6.0.1
python-multipart==0.0.6
httpx==0.26.0
aioping==0.4.0
